Add support for active partition detection through device ID.
This is required for BeagleBone, where the `mount` output is useless
(it shows `/dev/root`, which isn't even a device). We still call the
`mount` command, but this is only a guide to go straight to the right
file. If the file is found to be incorrect, we iterate the `/dev`
directory until we find a device matching the device ID of the root
directory.

Also improve and restructure tests a bit. Give each test a separate
test function, and test more of the error cases for unexpected
partition numbers. And a test for the added functionality, of course.

MEN-30