support/mender-device-identity: example platform integration script
The patch adds an example integration script for producing device
identity data. The script looks up a network interface with the lowest
ifindex, other than loopback interface, and grabs its MAC address. It is
assumed that /sys/class/net is available in the system.

The identity data is output to stdout and has a form of:
<key>=<value>

The script may produce some information to stderr, make sure that it's
accounted for by whoever is reading the output.

Signed-off-by: Maciej Borzecki <maciej.borzecki@open-rnd.pl>