mender: state handling
Extend state handling to accomodate states that the mender controller
can be in. The patch adds a TransitionState() method that may internally
transition controller from one state to another. In case of errors, the
controller reaches MenderStateError, recording the error that caused the
transition. The error can be obtained by calling LastError() method.

Signed-off-by: Maciej Borzecki <maciej.borzecki@open-rnd.pl>