Add deployment log functionality.
Deployment logs are organized as a hook to current logrus library.
There is a DeploymentLogManager which is main deployment logging
module allowing enabling and disabling logging and rotating log files.

Logs are organized as indiidual JSON entries written to text files.
The log fields are hard-coded and the current format looks like follows:
{"level":"info","msg":"test2","time":""}

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>