Update deployment logs sending tests.
As the structure of log message changed we need to update some tests.
Some tests are commented as it seems JSON validation is not working as
expected. Even though we are having the same JSON field, looks like order
of those is important as well.

TODO: add missing tests for DeploymentLogManager.GetLogs()

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>