Make sure to report failed update after rebooting with broken image.
If we are installing broken image device is rebooted reaching
`we have upgrade info but has flag is not set` point.
Here we are sending no information to the server that last update was
broken and we are entering infinite loop of fetching and installing
broken update.
Right now we will try to report broken update so that server will be able
to remove it from the queue.

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>