MEN-630 Check update status before sending report to server.
Previously when we've been not able to send status to the server after
successful update we've been rolling back to the last working image.
The problem was that `report state` has been saved as the last known state
with the status `success`. Once device was rebooted it continued to send
`success` to the server even though we rolled back.
This is fixing the issue by verifying exact state before sending report
to the server.
It also uses `reboot` state as a part of rollback.

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>