Revert "MEN-577: Add LMDB to Mender dependencies."
This reverts commit 822822c41eeccf88643caa91bc5db3eb3b95b871.

The dependency was originally added with the intention of using it to
replace file based client state persistence with one based on DB. This
has not happened yet and the code does not import
github.com/bmatsuo/lmdb-go. As a result, each time one uses `godep save`
godep will try to remove the library from the list of vendored
dependencies.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>