Add state scheduling mechanism.
This introduces simple scheduler waiting for various states to be run.
It is possible to add a state to be scheduled with given frequency.
It is also possible to remove the state from the scheduler.
The main part is the `Wait()` function which is calculating the time of
the next event to be run and starts the timer to be fired once the
scheduled event should be run.

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>