MEN-812 Do not store state as number.
So far we've been storing update state as number (`MenderState`).
There is potential risk in this method as after adding some more
states we might end up with some issues after upgrade.

Let's assume we are upgrading client from the version 1.0 to
version 1.1. In the version 1.1 some state was added which is changing
the order of states. In such a scenario, state stored in version 1.0
will be interpreted differently by version 1.1.

This is supposed to fix the ambiguity in storing and loading states.

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>