MEN-480: Make sure U-Boot environment update is atomic.
We need to update the environment in one operation, because otherwise
it can be in an in-between state if the device goes down unexpectedly.
The actual atomicity of the operation is ensured by U-Boot and its
redundant environment, we just need to call the right command.

One test case was removed because it's not possible to get such an
error anymore.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>