Do not enter error state after it is not possible to enable logger.
When update is in progress and we are rebooting device or we are
verifying update status, returning error state will cause device to
start the states sequence from the `init` state. This will prevent the
scenario where, after we can not start deployment logger, device state
is reseted and update procedure is stopped. Instead we will continue
update, but the logs will be missing.

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>