MEN-961 Make sure rollback happens after deployment is aborted
This is changing rollback state to perform reboot itself, instead of
switching to reboot state. The reason is we are storing state
information in reboot state and we are checking if deployment is valid.
In case of aborted deployment, which was correctly detected after
device was rebooted we've been entering rollback and then reboot state.
As deployment was aborted instead of continuing with reboot we
reported error and went back to error and then init state.
No rollback happened in this scenario.

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>