device: write data based on sector size
io.copy uses a hardcoded maximum of 32kb when performing the read/write
operations.

This is often not optimal and we should take in account what the
underlying devices belives it has as sector size.

One example where this is especially important is NAND based devices
which can only performe writes based on sector size and each write
generates an erase cycle and if writes are not sector size aligned it
will generate unnecessary wear on the NAND which has a limited amount of
erase cycles during its lifetime.

io.CopyBuffer is indentical to io.Copy with one difference, you can
provide a staging buffer which can be larger then 32kb. Lets pass
is a buffer that matches the sector size!

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@gmail.com>