block_device: add typeUBI property to structure
We need to check and cache if we are updating an UBI volume. We simply
check if inactive partition exists under /sys/class/ubi.

Also UBI block device path needs some special care due to the fact that the
Linux kernel does handle /dev/ubi0_0 as argument to root=.

Because of that we have only ubi0_0 as configuration parameters,
including /etc/mender/mender.conf and we need to fix that now that we
want to open the device and start writing to it.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@gmail.com>