ioctl: add fallback to check UBI block volumes
To avoid an special treatment of UBI .mender files, this is an attempt
to write generic methods to get sector size and device size.

It will first attempt to fetch the information using IOCTL requests,
which works on SD/MMC but will fail in UBI.

If it fails on IOCTL it will attempt to read UBI sysfs attributes of the
device, if it succeeds to read from sysfs we condsider it a success.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@gmail.com>