block_device: add imageSize in BlockDevice structure
This is preperation to adding UBI block device support.

Prior to writing an UBI volume one must say how much data one anticipate
to write. This is an security feature that makes sure to mark the UBI
volume as corrupt if not all data is written.

The IOCTL UBI_IOCVOLUP is used for this and must happen after the device
is opened. UBI will verify the data written when one closes the file
descriptor.

That is why we need the block_device to be aware of how much data it is
anticipating to write since it is the one that opens the device on first
write.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@gmail.com>