block_device: add call to UBI_IOCVOLUP prior to writing any data
UBI_IOCVOLUP takes "size" as an argument and that is the size one
anticipates to write. It is an security measure to make sure that all
data is written, otherwise UBI will mark the volume as corrupt making it
un-bootable instead of booting an "partial" image.

Changelog: Introduce experimental support for writing to UBI volumes

Signed-off-by: Mirza Krak <mirza.krak@gmail.com>