Add Transitions as a way to handle state scripts.
Each state will contain a transitions set indicating if state
script should be called or not. Depending if we are doing transition
Enter(), Leave() or Error() actions for the given state transition
will be called.

Instead of controller RunState() function we have TransitionState()
now taking additional parameter indicating the state to which transition
will happen. All the scripts executing and error handling logic is
placed inside this function.

Changelog: None

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>