MEN-1178: Fix misleading version being displayed for non-tagged builds.
Most builds would end up displaying something like
integration-0.1-abcdef because it happens to be on the master
branch. But this tag is ancient and not really descriptive of the
current version. So display pure commit hash instead when we are not
on an exact tag.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>
(cherry picked from commit 25d629de05368651a7ff8589e0d86a050d797070)