Let transition itself decide if it should ignore errors.
Some states needs to be handled even if scripts are returning errors.
Those states are Idle, Rollback, RollbackReboot and Error. Even if
some state scripts will error, we still need to enter those states.
For all error actions we are ignoring execution errors as well making
sure all error scripts will be executed.

Also in case of having a script which is not executable we will try
to execute others from the given state if have multiple.

Changelog: None

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>