Make sure that scripts we write are synced to disk before we return.
If we don't do this, and a power loss happens after we have flipped
the partitions (which *is* synced), but before the cache of these
files have been written, then the next boot would boot to the new
partition, but without the scripts. The behavior would then be wrong,
since we now boot the new update without honoring the scripts it came
with.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>