MEN-1511: Implement download resume feature.
This implements the ability of the client to resume download if the
connection is broken for any reason. It does this by replacing the
socket reader with a custom reader object that keeps track of the read
offset, and if it is broken at any point, the download is
reestablished at the same point and reading continues.

This requires some custom parsing of the 'Content-Range' header, since
Go does not have this support natively on the client side. We
primarily need the offset that the server gives us back, with some
surrounding code to check for sanity of the returned header. All
according to HTTP spec.

One weakness of this method is that it is not possible to abort the
update and have the client notice this in the middle of the download,
even if the connection is broken. But this would require changing
state to check for deployment status, which would be quite hard to do
while an update is in progress mid-stream. Not impossible, but a lot
of work, and it was deemed out of scope for this initial feature. It's
also no worse than it was before, since you cannot currently abort an
update mid-download either.

Changelog: Fix 'unexpected EOF' error when downloading large updates.

Changelog: Implement ability for client to resume a download from
where it left off if the connection is broken.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>