Add check for canary value in U-Boot environment.
The way it works is that we store the 'mender_check_saveenv_canary=1'
variable in the default U-Boot environment. We do not store the
corresponding 'mender_saveenv_canary=1' value in the default
environment, but instead we write this during the boot process. When
the Mender client finds the former variable, it will look for the
second value as well, and produce an error if it is not present. If
this happens it is an indication that the U-Boot boot loader and the
user space tools do not agree on the location of the environment. It
will also catch cases where mender_setup is not run.

Also reorganize the logging and error returning to include the error
when returning higher up in the callstack, otherwise the resulting log
entry is quite misleading.

Changelog: Add automatic check for canary value in U-Boot environment
to try to detect if there is a problem in the environment setup of
U-Boot and/or the u-boot-fw-utils tools.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>