Fall-back to environment and conf in active part probe
There are situations where it nearly impossible to find the active
root-filesystem based on probing mounted devices.

One example is this:

    /dev/root on /rom type squashfs (ro,relatime)

There is no information about the actual device device used in above
example.

One additional probe step could be parsing /proc/cmdline to find the
"root=" argument but this is essentially the same as looking at
"mender_boot_part" environment variable.

Hence we add a fall-back to only using the "mender_boot_part"
environment variable and comparing that to what we have in
"mender.conf" if we are not able to match against mounted devices.

Changelog: Fix active partition detection when using non-native
filesystems.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>