MEN-2073: Merge configuration settings from two files.
Changelog: Mender client now loads configuration settings from
both /etc/mender/mender.conf and (if it exists)
/var/lib/mender/mender.conf. The second file is located
on the data partition, so it allows any subset of configuration
changes to survive upgrades.

Previously, Mender loaded configuration settings from a
single file on the active root partition: /etc/mender/mender.conf
Like everything else on the root partition, this config file
is replaced with each upgrade.

Let us call /etc/mender/mender.conf the "main" configuration file.
Now there is an additional "fallback" configuration file,
/var/lib/mender/mender.conf, stored on the data partition.

If a setting is in the main file, it will be used.

If a setting is not in the main file but is in the fallback
file, the fallback value will be used.

If a setting is in both files, the value from the main
file will be used.

If both files exist but a setting does not appear in either
file, it will keep its defalut value ("" for a string or
0 for an integer).

If the main file does not exist but the fallback file exists,
settings from the fallback file will be used.

If the fallback file does not exist, but the main file exists,
settings from the main file will be used (this is the same
behavior as before this code change).

If neither file exists, an error occurs and Mender stops.

Use DEBUG severity to log final merged configuration settings.
We do not use INFO because that would expose sensitive
information (keys) by default.

Added --fallback-config command line option to override the default
location of the fallback configuration file.

Updated copyright years in config_test.go, partitions.go
to satisfy Travis CI validation.  (I don't understand
how partitions.go didn't cause a failure already.)

Signed-off-by: Don Cross <cosinekitty@gmail.com>