MEN-2285: device.go: Do "medium-sized" writes rather than individual sectors:
Testing showed that writing to the block device in increments of single
sectors (512 bytes on my eMMC device) contributed to an OOM condition in
the kernel when allocating DMA scatter-gather descriptors.

Rather than making a zillion small writes and expecting the kernel to
coalesce them for us when we're easily able put this data into already
(virtually) contiguous memory in userspace, we write to the block device
in "medium-size chunks" (around 1 MiB) that are a power-of-two multiple
of the sector size (obviating any concerns about sector misalignment
causing reduced performance or undue wear on the underlying storage).

The 1 MiB size could be made configurable if need be.

This commit introduces the following:

- Write to block device at some power-of-two multiple of the sector size.
- We introduce chunkedCopy() which ensures that writes happen exactly in
  increments of the specified size (except for the last write after EOF
  of the input).  Added unit test to verify correct behavior.

Changelog: Work around occasional OOM bug in mmc driver.
Signed-off-by: Jeremy Trimble <jeremy.trimble@twosixlabs.com>