module docs: Fix design flaw in streaming API regarding `streams-list`
We don't know the order of the streams upfront, so we cannot provide
the `streams-list` pipe originally envisioned in the spec. Instead,
provide a `stream-next` file which returns exactly one entry, the next
one, which is then meant to be repeatedly read until it returns empty.

Also, provide a few more convenience entries in the file API, and
clarify that the file API is subject to more entries in the future, as
part of API extensions.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>