MEN-2371: Write Update Module to do file(s) install
This update module allows a deployment of a single file or a full file
tree into an specific destintion dir in the device.

The info about the destination dir is contained in the meta-data while
the files themselves are bundled in a tarball sent in the Artifact.

Added the script to generate compatible Artifacts for the Update Module.

Changelog: Title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>