Fix race condition which mistook "reboot commenced" for "reboot finished".
The problem is that the reboot command returns success, and then
reboots. This will result in mender being killed by systemd, but there
is a window where the client can continue, and it will interpret this
as a successful reboot and continue with subsequent states.

Fix by waiting for the reboot to complete, and if it doesn't, return
error.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>