MEN-2392: Readd rootfs-image-v2 and make it user consumable.
Also clean it up and make it robust, configurable and atomic. Run a
diff against fe3a3836b40d99 if you want to know the differences.

Changelog: Add `rootfs-image-v2` as a demonstration of how to
reimplement Mender's `rootfs-image` update type as an update module.
It's also useful as inspiration if users want to make their own
slightly tweaked rootfs-image type.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>