Fix crash when rebooting to partition that is missing dual rootfs conf.
If the client is rebooted into a partition which is missing the dual
rootfs configuration, it would crash, and the error would persist for
as long as the database is present, and the config is not added. It is
difficult to recover from this situation, but we should at least try
to reboot, and in general, let the client run. This at least gives a
chance that you can fix the device with a different type of update.

For this we use a special stub installer, which fails at almost every
step, except for the rollback reboot. If it is unable to fix the
problem, the client will eventually break out of the state loop due to
the state counter, and report the failure to the server.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>