Update modules: Implement `NeedsArtifactReboot` -> `Automatic`.
The motivation behind is change is to promote the use of `Automatic`
in all update modules. We want this because in order to support
multiple payloads, an update module should not reboot the host in its
`ArtifactReboot` script. Doing so means that `ArtifactReboot` scripts
in later payloads do not get to execute. Instead it should leave this
to Mender using the `Automatic` mechanism. Multiple payloads have not
been implemented yet, but since `Automatic` needs to be supported by
the update module itself, we implement this part now so that we don't
accumulate too many modules without it before we get multiple payload
support.

Note that the most important test of this feature is by changing the
`rootfs-image-v2` script, which is then used by the integration test
which already exists.

Changelog: Title

MEN-2011

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>