file-install modules: Don't destroy original before we know we have a backup.
This was revealed by the `test_file_update_module` in integration,
which puts a file in the way of the directory. If it fails that early,
we should not delete the original, since we never created a backup.

More commonly, this problem would happen if there was no space for the
backup, which would obviously be disastrous. So also make sure we have
a *complete* backup.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>