Move Mender state machine to a producer-consumer pattern
NO-RUN-TESTS

This move will enable getting rid of all the storing of time, and
checking before and after, and weird handling of the special cases related
to null-time (events that should happen straight away). It is by no means
finished!! But more like a suggestion draft, as to whether I should follow this
up.

This will reduce the number of internal states significantly, and make
it easier to maintain (I think).

Feedback most welcome :)

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>