Fix UBI device size calculation
Prior to this change, the code in `getUbiDeviceSize` looked at the
`data_bytes` property of the UBI device/volume instead of looking at
the total size of the volume. This caused the size of a UBI device to
always get undereported and in some cases, such as a failed update,
could cause devices to no longer be able to get updated.

To demonstrate this, note that `Data bytes` is 2MB (after a 2MB file
is written to the volume) despite the total usable size being 50MB:

    root@unknown:~# dd if=/dev/urandom of=/tmp/blah bs=1M count=2
    2+0 records in
    2+0 records out
    root@unknown:~# ubiupdatevol /dev/ubi0_0 /tmp/blah
    root@unknown:~# ubinfo /dev/ubi0_0
    Volume ID:   0 (on ubi0)
    Type:        static
    Alignment:   1
    Size:        802 LEBs (52457216 bytes, 50.0 MiB)
    Data bytes:  2097152 bytes (2.0 MiB)
    State:       OK
    Name:        rootfsa
    Character device major/minor: 244:1

Then, trying to update `ubi0_0` with a ~29MB update using Mender
produced the error message:

    ERRO[0000] update (29093140 bytes) is larger than the size of device /dev/ubi0_0 (2097152 bytes)  module=device

This is a fairly critical issue as it effectively means that:

- You are never able to install an artifact that is bigger than your
current artifact
- If an update is interrupted and the UBI volume shows as 'corrupted',
the `Data Bytes` will read as 0 bytes which effectively stops ANY
update being applied.

The fix, implemented in this commit, is to calculate the size of the
device by multiplying the number of reserved erase blocks by the size
of each erase block.

changelog: title
Signed-off-by: Sam Lewis <sam.vr.lewis@gmail.com>