Enable the usage of the full Mender-Artifact version 3 format
With the updated Mender-Artifact version, the client can now support the entire
new Artifact version 3 format. This means that now the client can handle
provides, depends and augmented Artifacts through the use of the new
Mender-Artifact library.

Hence, remove the old sentinel which made sure this functionality was not
attempted through a `MissingFeatures` sentinel in the respective installer's
intialize methods.

The changes were needed in relation to migrating the 'mender-binary-delta'
update-module, ticket: MEN-2642. Since the client will need to pass depends and
provides information from Mender-Artifact to the Mender-binary-delta update
module.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>