MEN-2969: Fix: Store provide data across reboots in standalone mode
Previously the standalone mode would lose the provide data from the Artifact, as
the data was not stored, for when the device was rebooted.

Changelog: Now the client stores Artifact provides parameters across reboots in
standalone mode. Previously this data was ignored, and hence upgrading with an
Artifact with provides parameters these were lost.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>