Give the block device a file interface
This commit changes the interface for the block-device to abide by the regular
golang file interface. Hence, the block-device is given an open and a Write
method. Then only disrepancy is that the open-call requires a partition size
argument, in order to not overwrite the size of the given update.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>