MEN-2939: Optimize rootfs-update image writes
This commits adds a simple delta-writer to the writer stack. It works as
follows:

After receiving a chunk of bytes from the server, it reads data from the already
existing passive partition and compares the bytes in frames of N bytes. If the
bytes match, the write is simply skipped, and the next frame is considered. If
they don't match, the frame is considered dirty, and will be written.

This should help speed up the writing of rootfs-image updates, as in general
reading is a lot faster than writing. At least for SD-Cards.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>