No optimized writes on UBI are possible
Since any writes on UBI need to report prior to writing, how many bytes are to
be written, the optimized writes are not possible, due to the bytes discarded by
the optimized writer.

Therefore, the optimized writer is not added for UBI volume writes.

This commit therefore removes all the chained writers, and simply does a regular
write to the block-device filedescriptor.

Also, the opening of the block-device filedescriptor had to be moved to after
the size-checks of the block-device. As opening a UBI-block-device twice, will
result in a 'device or resource' busy error, and an update will be impossible.

Also added a couple of new debug prints to the block-device Open function.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>