MEN-3047: Fix crash when specified certificate can't be opened.
Changelog: Fix crash when specified certificate can't be opened.
Both the `ServerCertificate` setting and the system certificates are
now optional, in the sense that the client will run without them.
However, the client will not be able to connect without the right
certificates, so the main usecase of this change is to have a workable
client that will roll back if connections can't be made, instead of
exiting.

Also redid the `TestEmptySystemCertPool` test. This test actually
never worked as intended, because Go caches the system certificates
behind the scenes, so it only worked if you ran the test by itself.
The pass condition was not specific enough to catch this failure.
Reworked the test to work with an interface instead.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>