Move `--compression` flag inside `dump` and improve parsing.
It should be inside `dump`, because we don't know what other commands
we will be supporting inside `snapshot` and `--compression` may not be
appropriate for all of them.

Also make sure that we produce error on unknown flags and unit test
this. Tests of the actual implementation is in acceptance tests in
meta-mender.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>