utils/progress_bar: Move printer to a timed loop and simplify Tick
Tick is called for every iteration of the loop being tracked. There's no
point in printing on every iteration of the loop. Moreover, the snapshot
copy progressbar inhibits the progressbar from writing untill the output
fifo buffer (in case of pipe) is overfilled such that prompts from the
target process is not cluttered.

changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>