MEN-2587: Send Provides in the deployments API call
This enables the client to send provides data to the deployments endpoint during
an update check. The way it works is that the client will first do a POST to the
deployments/next endpoints, and then if the server is not Enterprise, a 404 will
be returned, upon which the client falls back to the standard GET
deployments/next functionality.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>