Rework freeze handler to use request/response model.
This was done to simplify the handler, now every interaction with the
handler is a request/response, so deadlocks are easy to analyze.

There is still one outstanding issue after this commit, but it appears
to be a bug in Go itself: SIGINT is not always delivered to sigChan.
It does not matter if you put the Notify call and the select on two
consecutive lines, it's random whether a ^C is delivered or not. Most
often it's not. I'm not able to reproduce it in a simple program, but
I cannot see how it could possibly be our fault if it doesn't even
work with Notify and select on consecutive lines. It only seems to
happen in cases where the writing is blocked due to a frozen
filesystem, so it may be related to the I/O thread that Go uses
somehow. If writing is proceeding as normal, then signal delivery
proceeds as normal as well, so it is possible to cancel an ongoing
operation. And a stalled operation due to a frozen filesystem will
also recover after the timer expires, so I'm not considering this a
big issue. It will just take a bit longer to cancel it.

[1] https://github.com/golang/go/issues/36420

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>