Fixed tests to go with logrus OOB
Previously the tests relied on accessing the writer of the mendersoftware/log
struct. Since this package does no longer exist, the tests had to be rewritten
partly, to now rely on the `sirupsen/hooks/test` functionality, which gives a
cleaner interface to test the logs with.

Also, this commit updates all the dates in the license headers.

As well as updating the LIC_FILES_CHKSUM file.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>