Fix Stat_t.Dev/Rdev type assumption
The unix.Major and unix.Minor functions takes an argument of type uint64 and the code currently passes in either stat.Rdev or stat.Dev into these. While stat.Dev and stat.Rdev are often defined as uint64 this is not so for every architecture. See for example mips64el where the fields instead are uint32 [1]. This causes a build failure (atleast with golang 1.14) on architectures where the assumtion doesn't hold true. This change simply casts the value to uint64 before passing it into unix.Major/Minor to fix the build problem.

[1]: https://sources.debian.org/src/golang-1.14/1.14-2/src/syscall/ztypes_linux_mips64le.go/#L107

Changelog: Title

Signed-off-by: Andreas Henriksson <andreas@fatal.se>