QA-153: Add Makefile to instrument the client for cover collection
This commit adds the ability to build the client with coverage functionality
enabled. That is it downloads the 'gobinarycoverage' tool, which instruments a
regular main package source code for binary coverage analysis. This adds a
function 'coverReport()' to the main file, which reports coverage for the binary
after execution.

However, in order for this to work, the exit paths in the binary need to call
'coverReport()' before exiting. Hence, the binary is patched with the patch file
found in 'patches/', which makes sure that 'coverReport()' is called before

1) The client calls reboot
2) The client exits normally

The target added is 'instrument-binary' and is a phony target, as it does not
build the binary, it is simply a transformational step, before calling 'go
build' like normal.

Note: This will change the source code in the directory, and hence all source
files are left 'dirty' after the 'instrument-binary' target has been run.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>