MEN-3115: Enable mTLS client authentication
This adds the ability to configure the client with a client certificate and
private key in order to enable mTLS in the client communication.

The implementation adds the configuration parameters to the Mender configuration
file, and issues an error, if a configuration is loaded in which a client
certificate is set, without a corresponding key. Likewise for the symmetric
case, in which the key is set, and the certificate is not. In short, either they
both have to be present, or missing.

The functionality is tested with a mocked mTLS server in which the client
attempts to upload an update log.

Changelog: MEN-3115: Add the ability to configure the client with a client certificate and
private key in order to enable mTLS in the client communication.

It also removes the `SkipVerify` option in the `HttpsClientConfig` in the Mender
configuration file, and adds a `SkipVerify` option to the top level of the
configuration file.

Changelog: Move the `SkipVerify` configuration option to the top-level of the
configuration file.