fix: Call the BN_ULONG macro explictly
This avoids having to manually cast to long, ulong, int, uint, or whatever
OpenSSL compiles to on your platform. According to the documentation:

https://www.openssl.org/docs/man1.1.1/man3/BN_set_word.html

'BN_ULONG is a macro that will be an unsigned integral type optimized for the
most efficient implementation on the local platform'.

Thus, doing this macro call is the intended way of typecasting, so that we don't
have to do manual #ifdefs

Changelog: None
Signed-off-by: Peter Grzybowski <peter@northern.tech>