Fix rootfs-image-v2 commit in standalone mode when upgrade fails
This changes the behavior of the ArtifactCommit function. Now, the variable
"upgrade_available" is verified, just like the built-in "rootfs-image",
and the commit fails if no upgrade is in progress. So, if an upgrade fails and
the user tries to commit after returning to the old OS, the commit will fail
and an appropriate message will be printed.

Changelog: Title

Signed-off-by: Pascal Hache <hacpa@touchtunes.com>